@@ -1,13 +1,11 @@
 package no.nav.pensjonsamhandling.maskinporten.validation.config
 
 import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.ConstructorBinding
 import org.springframework.boot.context.properties.source.InvalidConfigurationPropertyValueException
 import java.net.InetSocketAddress
 import java.net.Proxy
-import java.net.URL
+import java.net.URI
 
-@ConstructorBinding
 @ConfigurationProperties("maskinporten.validation")
 class MaskinportenValidatorProperties(
     environment: Environment.EnvType = Environment.EnvType.PROD,
@@ -20,14 +18,14 @@ class MaskinportenValidatorProperties(
         Environment.EnvType.VER2 -> Environment.Ver2
         Environment.EnvType.DEV -> Environment.Dev
         Environment.EnvType.CUSTOM -> Environment.Custom(
-            URL(
+            URI(
                 customUrl
                     ?: throw InvalidConfigurationPropertyValueException(
                         "maskinporten.validation.customUrl",
                         customUrl,
                         "Must be specified when environment is set."
                     )
-            )
+            ).toURL()
         )
     }
     val proxy = proxy?.run {
